Refacto ShoppingController - moved Body & Item entities into separate classes
